Detect winning combos via function call
GameModule.checkWinner(array, sign) takes an array and loops through
an array of possible indexes where that sign can be found. If any
are true, this means a win condition has been detected and the
function will return true.

This function still has yet to be implemented into the main program,
but functions perfectly well in the console. It will soon run
automtatically every turn when the player clicks a space to make
a valid turn.