Debug victor declared even if draw
The game was previously declaring a victor incorrectly even if a
draw was obtained. This was because there was nothing written to
tell the game what a draw was.

This was solved by an if else statement and an 'and' gate running
after victor detection is false. If and only if victory === false
AND there are no blank spaces left will not call the declareWinner
function and instead display that a draw occured.

Also debugged an issue where the game was changing turns even if
an invalid space was clicked. This was solved by moving the
function call to check winner and change turns inside of the
logic for detecting a valid space within the markSquare function.